+++ src/Game.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-Game::Game(unsigned int x, unsigned int y, string board, Location initialState, Location target, std::function<double(Location, Location)> heuristic) :
+Game::Game(unsigned int x, unsigned int y, string board, Location initialState, Location target, std::function<uint32_t(Location, Location)> heuristic) :
 			state(initialState), target(target), heuristic(heuristic){
 	a = 0;
 	if (board.length() != x * y) {
